app/requirements.txt
flask
uwsgi
pyotp
qrcode[pil]
markdown
bleach
itsdangerous
PyCryptodome
passlib
python-dotenv
flask_wtf